backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.0
python-multipart==0.0.6
pillow==10.1.0
opencv-python==4.8.1.78
transformers==4.35.2
# torch==2.1.1+cu121  # CUDA 12.1 version for GPU support - installed separately
accelerate==0.24.1
pdf2image==1.17.0
ollama==0.1.8
pydantic==2.5.0
pdfplumber==0.10.3
python-docx==1.1.0
nltk==3.8.1